+++ src/main/java/com/feeyo/redis/net/front/handler/PipelineCommandHandler.java
@@ -9,8 +9,8 @@ import org.slf4j.LoggerFactory;
 
 import com.feeyo.net.codec.redis.RedisRequest;
 import com.feeyo.net.codec.redis.RedisRequestType;
-import com.feeyo.net.codec.redis.RedisResponsePipelineDecoder;
-import com.feeyo.net.codec.redis.RedisResponsePipelineDecoder.PipelineResponse;
+import com.feeyo.net.codec.redis.RedisPipelineResponse;
+import com.feeyo.net.codec.redis.RedisPipelineResponseDecoderV2;
 import com.feeyo.net.nio.util.TimeUtil;
 import com.feeyo.redis.engine.manage.stat.StatUtil;
 import com.feeyo.redis.net.backend.BackendConnection;
@@ -50,13 +50,13 @@ public class PipelineCommandHandler extends AbstractPipelineCommandHandler {
 	// 
 	private class PipelineDirectTransTofrontCallBack extends DirectTransTofrontCallBack {
 
-		private RedisResponsePipelineDecoder decoder = new RedisResponsePipelineDecoder();
+		private RedisPipelineResponseDecoderV2 decoder = new RedisPipelineResponseDecoderV2();
 		
 		@Override
 		public void handleResponse(BackendConnection backendCon, byte[] byteBuff) throws IOException {
 
 			// 解析此次返回的数据条数
-			PipelineResponse pipelineResponse = decoder.decode( byteBuff );
+			RedisPipelineResponse pipelineResponse = decoder.decode( byteBuff );
 			if ( !pipelineResponse.isOK() )
 				return;
